feat : member entity 추가
