--- server/src/main/java/com/stackoverflow/stackoverflowclone/Member/entity/Member.java
@@ -1,35 +1,0 @@
-package com.stackoverflow.stackoverflowclone.Member.entity;
-
-import com.stackoverflow.stackoverflowclone.audit.Auditable;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.Column;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class Member extends Auditable {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "member_id")
-    private Long id;
-
-    private String nickname;
-
-    private String email;
-
-    private String password;
-
-    private String location;
-
-    private String bio_title;
-
-    private String bio_content;
-
-}
